MR21 fees calc

+++ src/routes/admin/IncompletePaymentsCard.tsx
@@ -295,24 +295,35 @@ console.log("Called send email");
       
       // remove from dependent tables
       // remove from enrollment
-      const { enrollmentDeleteerror } = await supabase
+      const { error: enrollmentDeleteerror } = await supabase
       .from('enrollment') 
       .delete()
       .eq('learner_id', learner_id);
       
       if (enrollmentDeleteerror) {
-        throw new Error('Failed to delete the record.');
+        throw new Error('Failed to delete the enrollment record.');
       }
-      
+      console.log("Deleted enrollment records for learner ", learner_id);
       // remove from learner
-      const { learnerDeleteError } = await supabase
+      // remove from payment
+      const { error: paymentDeleteerror } = await supabase
+      .from('payment') 
+      .delete()
+      .eq('learner_id', learner_id);
+      
+      if (paymentDeleteerror) {
+        throw new Error('Failed to delete the payment record.');
+      }
+      console.log("Deleted payment records for learner ", learner_id);
+      const { error: learnerDeleteError } = await supabase
         .from('Learner') 
         .delete()
         .eq('id', learner_id);
 
       if (learnerDeleteError) {
-        throw new Error('Failed to delete the record.');
+        throw new Error('Failed to delete the learner record.');
       }
+      console.log("Deleted learner records for learner ", learner_id);
     },
     onSuccess: () => {
       toast ({
